+++ src/apps/shared/commands/domain.py
@@ -5,4 +5,4 @@ class PatchRegister(InternalModel):
     file_path: str
     task_id: str
     description: str
-    order: int | None
+    order: int
