+++ src/apps/shared/commands/patches/__init__.py
@@ -1,1 +1,1 @@
-from apps.shared.commands.patches.sample import *
+from apps.shared.commands.patches.sample import *  # noqa: F401 F403
