+++ src/cli.py
@@ -8,8 +8,8 @@ os.chdir(dname)
 import typer  # noqa: E402
 
 from apps.answers.commands import convert_assessments  # noqa: E402
+from apps.shared.commands import patch  # noqa: E402
 from apps.workspaces.commands import arbitrary_server_cli  # noqa: E402
-from apps.shared.commands import patch
 
 cli = typer.Typer()
 cli.add_typer(arbitrary_server_cli, name="arbitrary")
